Shift task name and id serialization to TaskEvent
Fixes #269
